[CHARDET] DtlrExtractor.extract_one: add support for input_sizes
